sparsha-master/sparsha-master/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
torch
torchvision
numpy
requests
crewai
langchain-openai
python-dotenv
pydantic 